+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix an incorrect autocorrect for `Capybara/RSpec/PredicateMatcher` when `EnforcedStyle: explicit` and `expect` actual is a complex expression. ([@ydah])
 - Fix a false positive for `Capybara/SpecificActions` when `click` uses arguments or a block. ([@ydah])
 - Fix an incorrect autocorrect for `Capybara/SpecificFinders` when a selector class is used with nested `class:` options. ([@ydah])
+- Fix CSS attribute parsing when attribute values contain quotes. ([@ydah])
 
 ## 3.0.0 (2026-06-22)